Renamed grepAs to grepEach.
